Add more tests for conditions

+++ src/main/java/org/graphwalker/core/condition/RequirementCoverage.java
@@ -36,12 +36,12 @@ public final class RequirementCoverage implements StopCondition {
 
     private final double percent;
 
-    public RequirementCoverage(double percent) {
-        this.percent = percent/100;
+    public RequirementCoverage(int percent) {
+        this.percent = (double)percent/100;
     }
 
-    public double getPercent() {
-        return percent * 100;
+    public int getPercent() {
+        return (int)(percent * 100);
     }
 
     @Override
